format: Temporarily make pulsing colors static
Clients currently don't know how to handle pulsing colors, so until they
do, we'll make pulsing colors static for now.

The reason we're doing this for the time being is because the battery
module, when in a warning/alarm state, would appear dark (the dark part
of the pulse), not orange/red.